Extracted _BoostPythonEnum definition in a proper python file
