+++ pyrx/doc_utils/boost_meta.py
@@ -1,0 +1,18 @@
+from typing import Self, ClassVar, TypeVar
+
+T = TypeVar("T")
+
+
+class _BoostPythonEnumMeta(type):
+    # This is not a real class, it is just for better type hints
+
+    def __call__(cls: type[T], value: int) -> T: ...
+
+
+class _BoostPythonEnum(int, metaclass=_BoostPythonEnumMeta):
+    # This is not a real class, it is just for better type hints
+
+    values: ClassVar[dict[int, Self]]
+    names: ClassVar[dict[str, Self]]
+
+    name: str
